re-organize component in multiple files
